requirements.txt
flask==2.3.3
flask-cors==4.0.0
supabase==1.0.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
httpx==0.23.3
httpcore==0.16.3
python-multipart==0.0.6